Move session connection wrapper to session module
