+++ dbt/adapters/spark/connections.py
@@ -287,56 +287,6 @@ class PyodbcConnectionWrapper(PyhiveConnectionWrapper):
             self._cursor.execute(sql, *bindings)
 
 
-class SessionConnectionWrapper(object):
-    """Connection wrapper for the sessoin connection method."""
-
-    def __init__(self, handle):
-        self.handle = handle
-        self._cursor = None
-
-    def cursor(self):
-        self._cursor = self.handle.cursor()
-        return self
-
-    def cancel(self):
-        logger.debug("NotImplemented: cancel")
-
-    def close(self):
-        if self._cursor:
-            self._cursor.close()
-
-    def rollback(self, *args, **kwargs):
-        logger.debug("NotImplemented: rollback")
-
-    def fetchall(self):
-        return self._cursor.fetchall()
-
-    def execute(self, sql, bindings=None):
-        if sql.strip().endswith(";"):
-            sql = sql.strip()[:-1]
-
-        if bindings is None:
-            self._cursor.execute(sql)
-        else:
-            bindings = [self._fix_binding(binding) for binding in bindings]
-            self._cursor.execute(sql, *bindings)
-
-    @property
-    def description(self):
-        return self._cursor.description
-
-    @classmethod
-    def _fix_binding(cls, value):
-        """Convert complex datatypes to primitives that can be loaded by
-           the Spark driver"""
-        if isinstance(value, NUMBERS):
-            return float(value)
-        elif isinstance(value, datetime):
-            return f"'{value.strftime('%Y-%m-%d %H:%M:%S.%f')[:-3]}'"
-        else:
-            return f"'{value}'"
-
-
 class SparkConnectionManager(SQLConnectionManager):
     TYPE = 'spark'
 
@@ -507,7 +457,7 @@ class SparkConnectionManager(SQLConnectionManager):
                     conn = pyodbc.connect(connection_str, autocommit=True)
                     handle = PyodbcConnectionWrapper(conn)
                 elif creds.method == SparkConnectionMethod.SESSION:
-                    from .session import Connection  # noqa: F401
+                    from .session import Connection, SessionConnectionWrapper  # noqa: F401
                     handle = SessionConnectionWrapper(Connection())
                 else:
                     raise dbt.exceptions.DbtProfileError(
